add authentication, authorization

+++ OrchidFarmed.ProjectManagement.Application/CommandHandlers/DeleteProjectCommandHandler.cs
@@ -4,6 +4,7 @@ using MediatR;
 
 using OrchidFarmed.ProjectManagement.Application.Contracts.Commands;
 using OrchidFarmed.ProjectManagement.Domain.Repositories;
+using OrchidFarmed.ProjectManagement.Domain.Shared.Exceptions;
 
 namespace OrchidFarmed.ProjectManagement.Application.Commands;
 
@@ -21,6 +22,14 @@ public class DeleteProjectCommandHandler : IRequestHandler<DeleteProjectCommand>
     public async Task Handle(DeleteProjectCommand request, CancellationToken cancellationToken)
     {
         _validator.ValidateAndThrow(request);
+
+        var project = await _projectRepository.GetAsync(request.ProjectId);
+
+        if (project == null)
+            throw new ProjectNotFoundException();
+
+        if (project.UserId != request.UserId)
+            throw new ForbiddenOperationException();
 
         await _projectRepository.DeleteAsync(request.ProjectId);
